extending existing map instead of adding a new one
